Duplicated profile_screen and SVGPicture error has fixed
