+++ frontend/one_net/pubspec.yaml
@@ -19,7 +19,7 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ^3.6.1
+  sdk: ">=3.0.0 <4.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -33,7 +33,6 @@ dependencies:
     sdk: flutter
   flutter_svg: ^2.0.17
   provider: ^6.1.2
-  flutter_svg: ^1.0.3
   video_player: ^2.2.19
 
 dev_dependencies:
